フォーム部分とリスト部分を切り分ける
