+++ src/components/features/posts/postlist/style.scss
@@ -1,0 +1,46 @@
+.tasksBoard {
+  background-color: #fff;
+  margin-top: 20px;
+  padding-top: 10px;
+  padding-bottom: 10px;
+  padding-right: 40px;
+}
+
+p {
+  float: right;
+  text-align: right;
+  padding: 0;
+  color: #ff6161;
+  font-size: 1.2em;
+  cursor: pointer;
+}
+
+ul {
+  list-style: none;
+}
+
+li {
+  border-bottom: 1px solid #53bdff;
+  padding-bottom: 15px;
+  padding-top: 15px;
+  font-size: 1.3em;
+  color: #2f4f4f;
+
+  span {
+    background-color: #ff6161;
+    padding: 0px 3px 0px 3px;
+    color: #fff;
+    cursor: pointer;
+  }
+
+  input {
+    width: 13px;
+    height: 13px;
+  }
+}
+
+.checked {
+  text-decoration-line: line-through;
+  text-decoration-color: red;
+  color: red;
+}
